requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.2

# Authentication & Social Auth
PyJWT>=2.8.0
dj-rest-auth>=5.0.0
django-allauth>=0.59.0
google-auth>=2.23.0

# Database
psycopg2-binary>=2.9.9

# Admin UI
django-jazzmin>=2.6.0

# API Documentation
drf-spectacular>=0.27.0
django-extensions>=3.2.0
django-filter>=23.0

# Utilities
requests>=2.31.0
cryptography>=41.0.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.2.0

# Health Checks
django-health-check>=3.17.0

# Optional: Payments (commented out if not needed)
# stripe>=7.0.0
# dj-stripe>=2.10.3

# Optional: Async Tasks (commented out if not needed)
# celery>=5.3.0
# redis>=5.0.0
# django-celery-beat>=2.5.0

# Optional: Monitoring (commented out if not needed)
# sentry-sdk>=1.40.0